move all components to their own folder
