+++ assets/src/components/libraries/Book.jsx
@@ -4,8 +4,8 @@ import Paper from '@material-ui/core/Paper';
 import Button from '@material-ui/core/Button';
 import parse from 'url-parse';
 import BookDetail from './detail/BookDetail';
-import { borrowCopy, returnBook, joinWaitlist } from '../services/BookService';
-import BookModel from '../models/Book';
+import { borrowCopy, returnBook, joinWaitlist } from '../../services/BookService';
+import BookModel from '../../models/Book';
 
 
 const isWaitlistFeatureActive = () => {
