+++ assets/src/components/libraries/BookList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Book from './Book';
-import BookModel from '../models/Book';
+import BookModel from '../../models/Book';
 
 function BookList(props) {
   return (
