+++ assets/src/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import * as Sentry from '@sentry/browser';
 import ErrorBoundary from './utils/ErrorBoundary';
-import App from './App';
+import App from './components/App';
 
 import '../css/index.css';
 import '../css/App.css';
